+++ electron/preload.js
@@ -1,0 +1,6 @@
+const { contextBridge, ipcRenderer } = require("electron");
+
+contextBridge.exposeInMainWorld("amber", {
+  isElectron: true,
+  pickFolder: () => ipcRenderer.invoke("pick-folder"),
+});